# openapi.yaml
openapi: 3.0.0
info:
  title: Dokany API
  description: منصة دكاني - واجهات برمجة التطبيقات
  version: 1.1.0
  contact:
    name: Dokany Team
    url: https://dokany.com
servers:
  - url: https://api.dokany.com
    description: Production Server
  - url: http://localhost:3000
    description: Development Server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    storeIdHeader:
      type: apiKey
      in: header
      name: x-store-id
  schemas:
...
    MediaSignature:
      type: object
      properties:
        signature:
          type: string
        timestamp:
          type: integer
        cloudName:
          type: string
        apiKey:
          type: string
        publicId:
          type: string
        folder:
          type: string
...
security:
  - bearerAuth: []
  - storeIdHeader: []
paths:
...
  /media/cloudinary-sign:
    post:
      tags: [Media]
      summary: الحصول على توقيع Cloudinary للرفع
      x-implemented: true
      security:
        - bearerAuth: []
        - storeIdHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                publicId: { type: string }
                folder: { type: string }
                resourceType: { type: string, enum: [image, video, raw] }
                tags: { type: array, items: { type: string } }
      responses:
        '200':
          description: تم توليد التوقيع
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data:
                    $ref: '#/components/schemas/MediaSignature'
  /media/upload-url:
    post:
      tags: [Media]
      summary: الحصول على رابط رفع مباشر لـ B2
      x-implemented: true
      security:
        - bearerAuth: []
        - storeIdHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                fileName: { type: string }
                fileSize: { type: integer }
                mimeType: { type: string }
                type: { type: string, enum: [image, video] }
      responses:
        '200':
          description: تم تجهيز الرابط
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data:
                    type: object
                    properties:
                      uploadUrl: { type: string }
                      fileKey: { type: string }
                      bucketName: { type: string }
                      resourceType: { type: string }
  /media/confirm:
    post:
      tags: [Media]
      summary: تأكيد رفع الميديا
      x-implemented: true
      security:
        - bearerAuth: []
        - storeIdHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                url: { type: string }
                originalUrl: { type: string }
                type: { type: string, enum: [image, video, document] }
                mimeType: { type: string }
                filename: { type: string }
                size: { type: integer }
                productId: { type: string }
                categoryId: { type: string }
                isPrimary: { type: boolean }
                order: { type: integer }
      responses:
        '200':
          description: تم تأكيد الرفع
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data: { type: object }

  # ============================================================
  # 🔔 NOTIFICATIONS
  # ============================================================
...